+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to semantic versioning.
 - Improve `fromString` in `URLPath` to infallibly parse all strings.
 - Take a `URL` origin in `toURL` in `URLPath`, fixing origin retention.
 - Remove `fromStringO` and `toURLO` from `URLPath`.
+- Narrow input in `omit` in `ReadonlyRecord` to readonly.
 
 ## 0.19.1 (2024-04-08)